Split `tests/client/rpcNotifications`
