+++ tests/client/service/notificationsSend.test.ts
@@ -6,30 +6,25 @@ import os from 'os';
 import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
 import { Metadata } from '@grpc/grpc-js';
 import { DB } from '@matrixai/db';
-import { KeyManager, utils as keysUtils } from '@/keys';
-import { GRPCServer } from '@/grpc';
-import {
-  NodeConnectionManager,
-  NodeGraph,
-  NodeManager,
-  utils as nodesUtils,
-} from '@/nodes';
-import { Sigchain } from '@/sigchain';
-import { ForwardProxy, ReverseProxy } from '@/network';
-import {
-  NotificationsManager,
-  utils as notificationsUtils,
-} from '@/notifications';
-import { ACL } from '@/acl';
-import {
-  GRPCClientClient,
-  ClientServiceService,
-  utils as clientUtils,
-} from '@/client';
+import KeyManager from '@/keys/KeyManager';
+import GRPCServer from '@/grpc/GRPCServer';
+import NodeConnectionManager from '@/nodes/NodeConnectionManager';
+import NodeGraph from '@/nodes/NodeGraph';
+import NodeManager from '@/nodes/NodeManager';
+import Sigchain from '@/sigchain/Sigchain';
+import ForwardProxy from '@/network/ForwardProxy';
+import ReverseProxy from '@/network/ReverseProxy';
+import NotificationsManager from '@/notifications/NotificationsManager';
+import ACL from '@/acl/ACL';
+import GRPCClientClient from '@/client/GRPCClientClient';
 import notificationsSend from '@/client/service/notificationsSend';
+import { ClientServiceService } from '@/proto/js/polykey/v1/client_service_grpc_pb';
 import * as utilsPB from '@/proto/js/polykey/v1/utils/utils_pb';
 import * as notificationsPB from '@/proto/js/polykey/v1/notifications/notifications_pb';
-import { GRPCClientAgent } from '@/agent';
+import * as keysUtils from '@/keys/utils';
+import * as nodesUtils from '@/nodes/utils';
+import * as notificationsUtils from '@/notifications/utils';
+import * as clientUtils from '@/client/utils';
 import * as testUtils from '../../utils';
 
 describe('notificationsSend', () => {
@@ -57,8 +52,8 @@ describe('notificationsSend', () => {
         return 'signedNotification' as SignedNotification;
       });
     mockedSendNotification = jest
-      .spyOn(GRPCClientAgent.prototype, 'notificationsSend')
-      .mockResolvedValue(new notificationsPB.AgentNotification());
+      .spyOn(NodeConnectionManager.prototype, 'withConnF')
+      .mockImplementation();
   });
   afterAll(async () => {
     mockedGenerateKeyPair.mockRestore();
@@ -170,7 +165,7 @@ describe('notificationsSend', () => {
     grpcClient = await GRPCClientClient.createGRPCClientClient({
       nodeId: keyManager.getNodeId(),
       host: '127.0.0.1' as Host,
-      port: grpcServer.port,
+      port: grpcServer.getPort(),
       logger,
     });
   });
@@ -209,8 +204,7 @@ describe('notificationsSend', () => {
     expect(mockedSendNotification.mock.calls.length).toBe(1);
     expect(
       nodesUtils.encodeNodeId(mockedSendNotification.mock.calls[0][0]),
-    ).toBe(receiverNodeIdEncoded);
-    expect(mockedSendNotification.mock.calls[0][1]).toBe('signedNotification');
+    ).toEqual(receiverNodeIdEncoded);
     // Check notification content
     expect(mockedSignNotification.mock.calls[0][0]).toEqual({
       data: {
